Fixed a a couple bugs from before
Fixed a small bug where the logic got inverted do to addtion of the Disposable.
Also converted all Pointers ints to IntPtrs so that they are more garb colec (and thus C#) friendly

+++ NFC_CardReader/ACR122U/ACR122U_SmartCard.cs
@@ -21,7 +21,7 @@ namespace NFC_CardReader.ACR122U
         public ACR122U_SmartCard(WinSmartCard MakeFrom) : this(MakeFrom.Parent, MakeFrom.Card)
         {}
 
-        ACR122U_SmartCard(WinSmartCardContext Parent, int Card) : base(Parent, Card)
+        ACR122U_SmartCard(WinSmartCardContext Parent, IntPtr Card) : base(Parent, Card)
         {}
 
         #region DeviceUtilities
